# docs/openapi/firebird.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Firebird API",
    "description": "Control plane API for bare-metal servers, virtual machines, networking, storage, and identity management.",
    "version": "1.0",
    "contact": {
      "name": "API Support"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
...
  "basePath": "/api/v1",
  "schemes": [
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/compute/capacity": {
      "get": {
        "summary": "Get capacity summary by machine type",
        "description": "Returns the number of bare-metal servers per machine type, grouped by lifecycle state. Requires authentication.",
        "operationId": "MachineTypeService_GetCapacitySummary",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/v1GetCapacitySummaryResponse"
            }
          },
          "401": {
            "description": "Missing or invalid authentication token",
            "schema": {}
          }
        },
        "tags": [
          "Machine Types"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
...
    "/machine-types": {
      "get": {
        "summary": "List machine types",
        "description": "Lists all active machine types. Optionally filter by category (cpu or gpu).",
        "operationId": "MachineTypeService_ListMachineTypes",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/v1ListMachineTypesResponse"
            }
          }
        },
        "parameters": [
          {
            "name": "category",
            "description": "Optional category filter (cpu or gpu)",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "pageSize",
            "description": "Maximum number of items to return (default: 20, max: 100)",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32",
            "default": "20"
          },
          {
            "name": "pageToken",
            "description": "Token for fetching the next page of results",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "Machine Types"
        ]
      }
    },
    "/machine-types/{machineTypeId}": {
      "get": {
        "summary": "Get a machine type by ID",
        "description": "Retrieves the full hardware specification for a single machine type.",
        "operationId": "MachineTypeService_GetMachineType",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/v1MachineType"
            }
          },
          "404": {
            "description": "Machine type not found",
            "schema": {}
          }
        },
        "parameters": [
          {
            "name": "machineTypeId",
            "description": "Machine type resource ID (format: machine-type.ULID)",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "Machine Types"
        ]
      }
    },
...
  },
  "definitions": {
...
    "v1CapacitySummaryEntry": {
      "type": "object",
      "properties": {
        "machineType": {
          "description": "Machine type details",
          "allOf": [
            {
              "$ref": "#/definitions/v1MachineType"
            }
          ]
        },
        "total": {
          "type": "integer",
          "format": "int32",
          "description": "Total number of servers of this type"
        },
        "countByState": {
          "type": "object",
          "additionalProperties": {
            "type": "integer",
            "format": "int32"
          },
          "description": "Count per lifecycle state (e.g. available: 3, running: 5)"
        }
      }
    },
    "v1CpuSpec": {
      "type": "object",
      "properties": {
        "cores": {
          "type": "integer",
          "format": "int32",
          "description": "Physical CPU cores"
        },
        "threads": {
          "type": "integer",
          "format": "int32",
          "description": "Total CPU threads (including SMT/HT)"
        },
        "model": {
          "type": "string",
          "description": "CPU model name (e.g. AMD EPYC 9654)"
        },
        "architecture": {
          "type": "string",
          "description": "CPU architecture (x86_64, arm64)"
        }
      },
      "required": [
        "cores",
        "threads",
        "model",
        "architecture"
      ]
    },
...
    "v1GetCapacitySummaryResponse": {
      "type": "object",
      "properties": {
        "items": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/v1CapacitySummaryEntry"
          },
          "description": "Capacity entries per machine type"
        }
      }
    },
...
    "v1GpuSpec": {
      "type": "object",
      "properties": {
        "count": {
          "type": "integer",
          "format": "int32",
          "description": "Number of GPUs"
        },
        "model": {
          "type": "string",
          "description": "GPU model name (e.g. NVIDIA H100 SXM5 80GB)"
        },
        "memoryGb": {
          "type": "integer",
          "format": "int32",
          "description": "GPU memory per device in GB"
        }
      },
      "required": [
        "count",
        "model",
        "memoryGb"
      ]
    },
...
    "v1ListMachineTypesResponse": {
      "type": "object",
      "properties": {
        "items": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/v1MachineType"
          },
          "description": "List of machine types"
        },
        "nextPageToken": {
          "type": "string",
          "description": "Token for the next page, empty if no more results"
        }
      }
    },
...
    "v1MachineType": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "description": "Machine type resource ID (format: machine-type.ULID)"
        },
        "name": {
          "type": "string",
          "description": "Machine type name (e.g. bm-gpu-h100.8x)"
        },
        "displayName": {
          "type": "string",
          "description": "Human-readable display name"
        },
        "category": {
          "description": "Hardware category (CPU or GPU)",
          "allOf": [
            {
              "$ref": "#/definitions/v1MachineTypeCategory"
            }
          ]
        },
        "cpu": {
          "description": "CPU specification",
          "allOf": [
            {
              "$ref": "#/definitions/v1CpuSpec"
            }
          ]
        },
        "memoryGb": {
          "type": "integer",
          "format": "int32",
          "description": "Total memory in GB"
        },
        "gpu": {
          "description": "GPU specification (null for CPU-only types)",
          "allOf": [
            {
              "$ref": "#/definitions/v1GpuSpec"
            }
          ]
        },
        "storage": {
          "type": "string",
          "description": "Storage description (e.g. 2x 1.92TB NVMe SSD)"
        },
        "networkBandwidthGbps": {
          "type": "integer",
          "format": "int32",
          "description": "Network bandwidth in Gbps"
        },
        "description": {
          "type": "string",
          "description": "Optional longer description"
        },
        "isActive": {
          "type": "boolean",
          "description": "Whether this machine type is currently available for provisioning"
        },
        "createdAt": {
          "type": "string",
          "format": "date-time",
          "description": "Creation timestamp"
        },
        "updatedAt": {
          "type": "string",
          "format": "date-time",
          "description": "Last update timestamp"
        }
      },
      "required": [
        "id",
        "name",
        "displayName",
        "category",
        "cpu",
        "memoryGb",
        "createdAt",
        "updatedAt"
      ]
    },
    "v1MachineTypeCategory": {
      "type": "string",
      "enum": [
        "CPU",
        "GPU"
      ]
    },
...
  },
  "securityDefinitions": {
    "BearerAuth": {
      "type": "apiKey",
      "description": "Bearer token authentication (JWT). Format: 'Bearer \u003cyour-token\u003e'.",
      "name": "Authorization",
      "in": "header"
    },
    "OIDCClientCredentials": {
      "type": "oauth2",
      "description": "OIDC/OAuth2 client credentials grant to obtain a bearer access token.",
      "flow": "application",
      "tokenUrl": "/api/v1/auth/token"
    },
    "OIDCPassword": {
      "type": "oauth2",
      "description": "OIDC/OAuth2 password grant (ROPC) to obtain a bearer access token.",
      "flow": "password",
      "tokenUrl": "/api/v1/auth/token"
    }
  },
  "security": [
    {
      "BearerAuth": []
    },
    {
      "OIDCPassword": []
    },
    {
      "OIDCClientCredentials": []
    }
  ]
}